Updated versions of sensor drivers
