+++ cfg/tools/capture_images.yaml
@@ -1,4 +1,4 @@
-output_dir: /nfs/diskstation/projects/dex-net/objects/physical
+output_dir: /nfs/diskstation/projects/dex-net/segmentation/datasets/mars
 calib_dir: /nfs/diskstation/calib
 
 sensors:
