+++ perception/constants.py
@@ -5,7 +5,7 @@ METERS_TO_MM = 1000.0
 MM_TO_METERS = 1.0 / METERS_TO_MM
 
 COLOR_IMAGE_EXTS = ['.png', '.jpg']
-MIN_DEPTH = 0.0
+MIN_DEPTH = 0.5
 MAX_DEPTH = 1.0
 MAX_IR = 65535
