+++ perception/rgbd_sensors.py
@@ -2,7 +2,7 @@
 RGBD Sensor factory
 Author: Jeff Mahler
 """
-from . import Kinect2Sensor, VirtualKinect2Sensor, PrimesenseSensor, VirtualPrimesenseSensor, PrimesenseSensor_ROS, EnsensoSensor, PhoXiSensor
+from . import Kinect2Sensor, PrimesenseSensor, VirtualSensor, PrimesenseSensor_ROS, EnsensoSensor, PhoXiSensor
 
 class RgbdSensorFactory:
     """ Factory class for Rgbd camera sensors. """
@@ -23,9 +23,6 @@ class RgbdSensorFactory:
             s = Kinect2Sensor(packet_pipeline_mode=cfg['pipeline_mode'],
                               device_num=cfg['device_num'],
                               frame=cfg['frame'])
-        elif sensor_type == 'virtual_kinect2':
-            s = VirtualKinect2Sensor(cfg['image_dir'],
-                                     frame=cfg['frame'])
         elif sensor_type == 'primesense':
             flip_images = True
             if 'flip_images' in cfg.keys():
@@ -33,9 +30,9 @@ class RgbdSensorFactory:
             s = PrimesenseSensor(auto_white_balance=cfg['auto_white_balance'],
                                  flip_images=flip_images,
                                  frame=cfg['frame'])
-        elif sensor_type == 'virtual_primesense':
-            s = VirtualPrimesenseSensor(cfg['image_dir'],
-                                        frame=cfg['frame'])
+        elif sensor_type == 'virtual':
+            s = VirtualSensor(cfg['image_dir'],
+                              frame=cfg['frame'])
         elif sensor_type == 'primesense_ros':
             s = PrimesenseSensor_ROS(frame=cfg['frame'])
         elif sensor_type == 'ensenso':
